+++ src/main/java/org/findy/findy_be/auth/oauth/handler/OAuth2AuthenticationSuccessHandler.java
@@ -16,8 +16,8 @@ import org.findy.findy_be.auth.oauth.token.AuthToken;
 import org.findy.findy_be.auth.oauth.token.AuthTokenProvider;
 import org.findy.findy_be.common.config.AppProperties;
 import org.findy.findy_be.common.utils.CookieUtil;
-import org.findy.findy_be.user.entity.RoleType;
-import org.findy.findy_be.user.entity.UserRefreshToken;
+import org.findy.findy_be.user.domain.RoleType;
+import org.findy.findy_be.user.domain.UserRefreshToken;
 import org.findy.findy_be.user.repository.UserRefreshTokenRepository;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
