+++ src/main/java/org/findy/findy_be/auth/oauth/service/CustomOAuth2UserService.java
@@ -8,8 +8,8 @@ import org.findy.findy_be.auth.oauth.info.OAuth2UserInfo;
 import org.findy.findy_be.auth.oauth.info.OAuth2UserInfoFactory;
 import org.findy.findy_be.bookmark.application.BookmarkService;
 import org.findy.findy_be.common.exception.custom.OAuthProviderMissMatchException;
-import org.findy.findy_be.user.entity.RoleType;
-import org.findy.findy_be.user.entity.User;
+import org.findy.findy_be.user.domain.RoleType;
+import org.findy.findy_be.user.domain.User;
 import org.findy.findy_be.user.repository.UserRepository;
 import org.springframework.security.authentication.InternalAuthenticationServiceException;
 import org.springframework.security.core.AuthenticationException;
