+++ src/main/java/org/findy/findy_be/auth/oauth/service/CustomUserDetailsService.java
@@ -3,7 +3,7 @@ package org.findy.findy_be.auth.oauth.service;
 import static org.findy.findy_be.common.exception.ErrorCode.*;
 
 import org.findy.findy_be.auth.oauth.domain.UserPrincipal;
-import org.findy.findy_be.user.entity.User;
+import org.findy.findy_be.user.domain.User;
 import org.findy.findy_be.user.repository.UserRepository;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
