+++ src/main/java/org/findy/findy_be/bookmark/domain/Bookmark.java
@@ -5,7 +5,7 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import org.findy.findy_be.common.entity.BaseEntity;
-import org.findy.findy_be.user.entity.User;
+import org.findy.findy_be.user.domain.User;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
