+++ src/main/java/org/findy/findy_be/user/application/UserService.java
@@ -1,6 +1,6 @@
 package org.findy.findy_be.user.application;
 
-import org.findy.findy_be.user.entity.User;
+import org.findy.findy_be.user.domain.User;
 import org.findy.findy_be.user.repository.UserRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
