+++ src/main/java/org/findy/findy_be/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package org.findy.findy_be.user.repository;
 
-import org.findy.findy_be.user.entity.User;
+import org.findy.findy_be.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
